requirements.txt
treamlit==1.52.2
pandas==3.0.0 
numpy==2.4.1 
mysql-connector-python>=8.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
mysql-connector-python>=8.1.0